11. [설정] securityConfig 경로 수정 & package 변경
