+++ src/main/java/platform/ads/api/v1/hello/controller/HelloController.java
@@ -1,4 +1,4 @@
-package platform.ads.api.v2.controller;
+package platform.ads.api.v1.hello.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -6,18 +6,18 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import platform.ads.api.v2.service.HelloV2Service;
 import platform.ads.config.base.Api;
 import platform.ads.config.base.ResultCode;
 import platform.ads.config.base.ResultVo;
+import platform.ads.api.v1.hello.service.HelloService;
 
 @Slf4j
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/v2")
-public class HelloV2Controller {
+@RequestMapping("/api/v1")
+public class HelloController {
 
-    private final HelloV2Service helloV2Service;
+    private final HelloService helloService;
 
     @GetMapping("/hello")
     public ResponseEntity<Api<ResultVo>> hello() {
@@ -25,7 +25,7 @@ public class HelloV2Controller {
             Api.<ResultVo>builder()
                .code(ResultCode.SUCCESS.CODE)
                .message(ResultCode.SUCCESS.MESSAGE)
-               .data(helloV2Service.hello())
+               .data(helloService.hello())
                .build()
         );
     }
