+++ src/main/java/platform/ads/api/v1/hello/service/HelloService.java
@@ -1,4 +1,4 @@
-package platform.ads.api.v1.service;
+package platform.ads.api.v1.hello.service;
 
 import org.springframework.stereotype.Service;
 import platform.ads.config.base.ResultVo;
