+++ src/main/java/platform/ads/config/security/SecurityConfig.java
@@ -18,7 +18,7 @@ import org.springframework.security.web.SecurityFilterChain;
 @EnableWebSecurity
 public class SecurityConfig {
 
-    private static final String WHITE_LIST = "/api/v1/auth/****";
+    private static final String WHITE_LIST = "/api/v1/member/*auth*";
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
@@ -32,9 +32,12 @@ public class SecurityConfig {
                     .requestMatchers(PathRequest.toStaticResources().atCommonLocations())
                     .permitAll()
                     .requestMatchers(
-                        antMatcher("/swagger-ui/**"),
-                        antMatcher("/swagger-resources/*"),
+                        antMatcher("/swagger-ui/**"),       // swagger-ui
+                        antMatcher("/swagger-resources/*"),     // swagger-resources
                         antMatcher("/v3/api-docs/**")
+                    ).permitAll()
+                    .requestMatchers(
+                        antMatcher("/api/*/hello/**")   // hello test
                     ).permitAll()
                     .requestMatchers(antMatcher(WHITE_LIST)).permitAll()
                     .anyRequest().authenticated()
